support for playing episodes by number
reworks show.py so that vlcExe is a class member
instead of being passed to every 'play' method.
(still an optional parameter to those methods)

changes the existing playEpisode function to
'private' and adds a new playEpisode function
that takes as arguments a season and episode
number.

Also changes the whitespace from tabs to spaces.